Splitting the code
